菜单事件响应的支持&自定义菜单的支持

+++ src/main/java/top/jinhaoplus/wechathelper/wechat/message/response/TemplateMsgSendResponse.java
@@ -5,13 +5,13 @@ import top.jinhaoplus.wechathelper.wechat.api.response.APIResponse;
 
 public class TemplateMsgSendResponse extends APIResponse {
 
-    private Integer msgid;
+    private Long msgid;
 
-    public Integer getMsgid() {
+    public Long getMsgid() {
         return msgid;
     }
 
-    public void setMsgid(Integer msgid) {
+    public void setMsgid(Long msgid) {
         this.msgid = msgid;
     }
 }
